@@ -14,7 +14,7 @@ description: A simple comic viewer application
 version: 1.4.9+37
 
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: ">=2.7.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -36,19 +36,19 @@ dependencies:
   yin_drag_sacle: ^0.0.2
   flutter_swiper: ^1.1.6
   event_bus: ^1.1.1
-  permission_handler: ^5.0.1
+  permission_handler: ^6.1.3
   path_provider: ^2.0.0
   provider: ^4.3.2+2
   toast: ^0.1.5
   uni_links: ^0.4.0
   share: ^0.6.5+2
-  file_picker: ^2.1.5
+  file_picker: ^3.0.1
   flutter_downloader: ^1.5.2
   logger: ^0.9.4
   flutter_widget_from_html: ^0.6.0
   flutter_easyrefresh: ^2.1.8
-  connectivity: ^2.0.2
-  battery: ^1.0.7
+  connectivity: ^3.0.3
+  battery: ^2.0.1
   flutter_spinkit: "^4.1.2"
   version: ^1.2.0
   direct_select_flutter: ^1.0.7
